backend/requirements.txt
fastapi[standard]
uvicorn
python-multipart
pillow
albumentations
opencv-python-headless
numpy
ibm-watsonx-ai
google-genai
